% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omneohealth.R
\docType{package}
\name{omneohealth}
\alias{omneohealth}
\alias{omneohealth-package}
\title{The Oxford Maternal and Neonatal Health study aims to synthesise the current
evidence on maternal and neonatal health in various countries through
systematic collection of related secondary data from multiple sources. Then
based on this synthesis, gaps in knowledge and corresponding data are
identified in order to inform researchers on possible future areas of study.
This package brings together all the analytic functions created and used for
this study.}
\description{
The Oxford Maternal and Neonatal Health study aims to synthesise the
    current evidence on maternal and neonatal health in various countries through
    systematic collection of related secondary data from multiple sources. Then
    based on this synthesis, gaps in knowledge and corresponding data are identified
    in order to inform researchers on possible future areas of study. This package
    brings together all the analytic functions created and used for this study.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernestgmd@gmail.com} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Proochista Ariana \email{proochista@gmail.com} [copyright holder]
  \item Sophie Janet [copyright holder]
}

}
\keyword{internal}
